requirements.txt
gooey
argparse
srt
python-dateutil